+++ authpass/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add `CONTRIBUTORS.md` to about dialog. #237 (@ATofighi)
 * Fixed potential crash in Preferences screen.
+* Fixed broken 'manage groups'.
 
 # 1.8.1 - 2021-07-13