+++ authpass/CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add links to forum, report error to forum https://forum.authpass.app/
 * WebDAV: Improve file not found error message.
 * Fix bug when changing database name.
+* Added Dutch translations.
 
 # 1.8.1 - 2021-07-13