+++ authpass/CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.8.2 - unreleased
+# 1.8.2 - 2021-07-22
 
 * Add `CONTRIBUTORS.md` to about dialog. #237 (@ATofighi)
 * Fixed potential crash in Preferences screen.